Split normalization internals into _normalize.py (#61)
